restructure nodes and actions
