+++ webui/react/src/omnibar/trees/dev.ts
@@ -1,7 +1,7 @@
 import { setServerAddress } from 'dev';
 import { globalStorage } from 'globalStorage';
 import { userPreferencesStorage } from 'hooks/useStorage';
-import { alertAction } from 'omnibar/actions';
+import { alertAction } from 'omnibar/trees/actions';
 import { Children, TreeNode } from 'omnibar/types';
 import { serverAddress } from 'routes/utils';
